Fix incorrect parameter checks inside searchCatalogItemsRequest
keywords, identifiers, and marketplace_ids can each be comma-separated strings that have a limit on
values. The code was trying to use count() to check for that, which doesn't work. I added explode(",", $var)
to fix this.

Imo, it would make more sense if these parameters were typed as arrays to begin with, but making a change like
that would break existing implementations. This method was chosen so it will "just work" for anyone's existing
usage of the library.