Build improvements
This adds a `build` and `clean` commands to `composer.json` to enable
easier discovery of the build process for new developers.

Due to the `clean` deleting the raw directory. I've moved the gitignore
into the top level `.gitignore`.

The available commands are added to the README.md

Add single schema generation commands to README

We need to explain about the need for `--` in composer commands