Update requests with a Json body
Previously we were assuming that any POST or PATCH request would have a
Json body, but this had many false positives and false negatives.

This commit builds on https://github.com/highsidelabs/saloon-sdk-generator/pull/3
to look for request bodies with the `application/json` content type and
is a PUT, POST or PATCH request.

As a result, some requests have the trait removed and others have it
added.